Address PR review - rename to form-completion-status
