+++ distribution/src/main/provisioning/50-lfs.txt
@@ -27,7 +27,7 @@
     # Customizes the page displayed during startup, before all the bundles are activated
     ca.sickkids.ccm.lfs/lfs-startup-customization/${lfs.version}
 
-    ca.sickkids.ccm.lfs/lfs-answer-flags/${lfs.version}
+    ca.sickkids.ccm.lfs/lfs-form-completion-status/${lfs.version}
 
     # Add support for external authenticators
     ca.sickkids.ccm.lfs/lfs-ldap-support/${lfs.version}/slingfeature
