+++ modules/form-completion-status/pom.xml
@@ -25,7 +25,7 @@
     <version>0.1-SNAPSHOT</version>
   </parent>
 
-  <artifactId>lfs-answer-flags</artifactId>
+  <artifactId>lfs-form-completion-status</artifactId>
   <packaging>bundle</packaging>
   <name>Flags form answers as incomplete or invalid</name>
