+++ modules/pom.xml
@@ -32,7 +32,7 @@
 
   <modules>
     <module>startup-customization</module>
-    <module>answer-flags</module>
+    <module>form-completion-status</module>
     <module>custom-jcr-oak-server</module>
     <module>commons</module>
     <module>ui-extension</module>
